chatGPT-socket-server/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
flask-socketio==5.3.4
eventlet==0.33.3
requests==2.31.0
openai==0.27.4
python-dotenv==0.19.2
